backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
watchfiles==0.22.0
websockets==12.0